+++ database/migrations/2026_04_09_121330_add_main_image_to_products_table.php
@@ -11,9 +11,14 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::table('products', function (Blueprint $table) {
-            $table->string('main_image')->nullable()->after('image');
-        });
+        if (!Schema::hasColumn('products', 'main_image')) {
+
+            Schema::table('products', function (Blueprint $table) {
+                $table->string('main_image')
+                    ->nullable()
+                    ->after('image');
+            });
+        }
     }
 
     /**
@@ -21,8 +26,11 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('products', function (Blueprint $table) {
-            $table->dropColumn('main_image');
-        });
+        if (Schema::hasColumn('products', 'main_image')) {
+
+            Schema::table('products', function (Blueprint $table) {
+                $table->dropColumn('main_image');
+            });
+        }
     }
 };
